RF ManagePluginsPage,create ManagePluginsAdvancedPage, add_FindBy
