+++ src/test/java/school/redrover/ManagePluginsTest.java
@@ -19,16 +19,5 @@ public class ManagePluginsTest extends BaseTest {
 
         Assert.assertEquals(actualListOfTasks, expectedListOfTasks);
     }
-
-    @Test
-    public void testServerInfoBoxAdvancedSettingsPluginPage(){
-        final String expectedInfoServerText = "If your Jenkins server sits behind a firewall ";
-        String ServerInfoBox = new MainPage(getDriver())
-                .clickManageJenkinsPage()
-                .clickManagePlugins()
-                .clickAdvancedSettings()
-                .clickExtraInfoServerIcon()
-                .getExtraInfoServerTextBox();
-        Assert.assertTrue(ServerInfoBox.contains(expectedInfoServerText));
-    }
 }
+
